move plate block to softmax pro structure
